separate types into own files
