+++ user/user.go
@@ -11,66 +11,16 @@ import (
 	"strings"
 	"time"
 
-	"github.com/alyjay/xiv-be/character"
 	database "github.com/alyjay/xiv-be/database"
+	types "github.com/alyjay/xiv-be/types"
 	"github.com/google/uuid"
 
 	_ "github.com/lib/pq"
 )
 
-type User struct {
-	ID          string                `json:"id,omitempty" db:"id"`
-	Username    string                `json:"username" db:"username"`
-	DiscordId   string                `json:"discord_id" db:"discord_id"`
-	Avatar      string                `json:"avatar" db:"avatar"`
-	AccentColor string                `json:"accent_color" db:"accent_color"`
-	Characters  []character.Character `json:"characters"`
-	AuthToken   string                `json:"auth_token" db:"auth_token,omitempty"`
-	Expires     int64                 `json:"expires" db:"expires,omitempty"`
-}
-
-type AuthTokenRequest struct {
-	ClientId     string `json:"client_id"`
-	ClientSecret string `json:"client_secret"`
-	Code         string `json:"code"`
-	GrantType    string `json:"grant_type"`
-	RedirectURI  string `json:"redirect_uri"`
-	Scope        string `json:"scope"`
-}
-
-type AuthTokenResponse struct {
-	AccessToken  string `json:"access_token"`
-	TokenType    string `json:"token_type"`
-	ExpiresIn    int    `json:"expires_in"`
-	RefreshToken string `json:"refresh_token"`
-	Scope        string `json:"scope"`
-}
-
-type DiscordUserResponse struct {
-	Id                   string `json:"id"`
-	Username             string `json:"username"`
-	Avatar               string `json:"avatar"`
-	Discriminator        string `json:"discriminator"`
-	PublicFlags          int    `json:"public_flags"`
-	Flags                int    `json:"flags"`
-	Banner               string `json:"banner"`
-	AccentColor          int    `json:"accent_color"`
-	GlobalName           string `json:"global_name"`
-	AvatarDecorationData map[string]struct {
-		SkuId string `json:"sku_id"`
-		Asset string `json:"asset"`
-	} `json:"avatar_decoration_data"`
-	BannerColor  string              `json:"banner_color"`
-	Clan         map[string]struct{} `json:"clan"`
-	PrimaryGuild map[string]struct{} `json:"primary_guild"`
-	MfaEnabled   bool                `json:"mfa_enabled"`
-	Locale       string              `json:"locale"`
-	PremiumType  int                 `json:"premium_type"`
-}
-
 func GetUser(w http.ResponseWriter, r *http.Request) {
-	var user User
-	var characters []character.Character
+	var user types.User
+	var characters []types.Character
 	id := r.URL.Query().Get("id")
 
 	db, err := database.GetDb()
@@ -96,7 +46,7 @@ func GetUser(w http.ResponseWriter, r *http.Request) {
 	json.NewEncoder(w).Encode(user)
 }
 
-func GetAccentColor(user *DiscordUserResponse) (color string) {
+func GetAccentColor(user *types.DiscordUserResponse) (color string) {
 	if user.AccentColor != 0 {
 		color = fmt.Sprintf("#%x", user.AccentColor)
 	} else if user.BannerColor != "" {
@@ -112,7 +62,7 @@ func LoginUser(w http.ResponseWriter, r *http.Request) {
 	site_url, _ := os.LookupEnv("SITE_URL")
 	code := r.URL.Query().Get("code")
 	pUrl := "https://discord.com/api/oauth2/token"
-	b := AuthTokenRequest{
+	b := types.AuthTokenRequest{
 		ClientId:     client_id,
 		ClientSecret: client_secret,
 		Code:         code,
@@ -144,7 +94,7 @@ func LoginUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var result AuthTokenResponse
+	var result types.AuthTokenResponse
 	err = json.Unmarshal([]byte(body), &result)
 	if err != nil {
 		http.Redirect(w, r, site_url+"/error", http.StatusInternalServerError)
@@ -162,7 +112,7 @@ func LoginUser(w http.ResponseWriter, r *http.Request) {
 
 	defer res.Body.Close()
 
-	var userResult DiscordUserResponse
+	var userResult types.DiscordUserResponse
 	_ = json.NewDecoder(res.Body).Decode(&userResult)
 
 	newUUID := uuid.NewString()
@@ -174,7 +124,7 @@ func LoginUser(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	var user User
+	var user types.User
 	user.ID = newUUID
 	err = db.Get(&user, `SELECT * FROM users WHERE discord_id=$1`, userResult.Id)
 	if err != nil {
